Fix: Force execution of workspace sync patch by renaming file
This commit provides the definitive fix for the missing 'Scale Barcode Settings' shortcut.

The root cause of the issue was that Frappe was not executing the patch script because it had the same filename as previous, failed attempts. The Frappe patching system keeps a record of executed patches and was skipping the file.

This fix resolves the issue by:
1. Renaming the patch file from `add_scale_barcode_settings_link_to_workspace.py` to `resync_pos_awesome_workspace.py`. This forces Frappe to see it as a new, un-executed patch.
2. Updating `patches.txt` to point to the new patch name.
3. Ensuring the new patch contains the most robust "Force Sync" logic, which overwrites the workspace with the JSON definition and resets the `custom` flag.

This will ensure the patch runs during the next `bench migrate` and correctly updates the workspace for all users.
