+++ posawesome/patches.txt
@@ -6,4 +6,4 @@ posawesome.patches.add_pos_opening_shift_to_pos_invoice
 posawesome.patches.add_pos_invoice_field_to_sales_invoice_reference
 posawesome.patches.add_sales_person_filter_to_pos_profile
 posawesome.patches.add_promotional_scheme_link_to_workspace
-posawesome.patches.add_scale_barcode_settings_link_to_workspace
+posawesome.patches.resync_pos_awesome_workspace
